+++ proto/control-api/src/control/member.rs
@@ -4,26 +4,36 @@ use std::{collections::HashMap, fmt, str::FromStr, time::Duration};
 
 use derive_more::{AsRef, Display, Error, From, Into};
 use ref_cast::RefCast;
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 use url::Url;
 
-use super::{endpoint, room, Endpoint};
+use super::{endpoint, room, Pipeline};
 
 /// Media [`Element`] representing a client authorized to participate in some
 /// bigger media pipeline ([`Room`], for example).
 ///
 /// [`Element`]: crate::Element
 /// [`Room`]: crate::Room
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Member {
     /// ID of this [`Member`] media [`Element`].
     ///
     /// [`Element`]: crate::Element
     pub id: Id,
 
+    /// [`Member`] spec.
+    pub spec: Spec,
+}
+
+/// [`Member`] spec.
+#[derive(Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+pub struct Spec {
     /// Media pipeline representing this [`Member`] media [`Element`].
     ///
     /// [`Element`]: crate::Element
-    pub pipeline: HashMap<endpoint::Id, Endpoint>,
+    pub spec: Pipeline<endpoint::Id, endpoint::Spec>,
 
     /// [`Credentials`] to authenticate this [`Member`] in [Client API] with.
     ///
@@ -50,6 +60,7 @@ pub struct Member {
     /// Once reached, this [`Member`] is considered being idle.
     ///
     /// [Client API]: https://tinyurl.com/266y74tf
+    #[cfg_attr(feature = "serde", serde(default, with = "humantime_serde"))]
     pub idle_timeout: Option<Duration>,
 
     /// Timeout of reconnecting for this [`Member`] via [Client API].
@@ -57,6 +68,7 @@ pub struct Member {
     /// Once reached, this [`Member`] is considered disconnected.
     ///
     /// [Client API]: https://tinyurl.com/266y74tf
+    #[cfg_attr(feature = "serde", serde(default, with = "humantime_serde"))]
     pub reconnect_timeout: Option<Duration>,
 
     /// Interval of pinging with heartbeat messages this [`Member`] via
@@ -66,6 +78,7 @@ pub struct Member {
     /// configured.
     ///
     /// [Client API]: https://tinyurl.com/266y74tf
+    #[cfg_attr(feature = "serde", serde(default, with = "humantime_serde"))]
     pub ping_interval: Option<Duration>,
 }
 
@@ -86,6 +99,8 @@ pub struct Member {
     PartialOrd,
     RefCast,
 )]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(transparent))]
 #[from(types(String))]
 #[into(owned(types(String)))]
 #[repr(transparent)]
@@ -95,6 +110,13 @@ pub struct Id(Box<str>);
 impl<'a> From<&'a str> for Id {
     fn from(s: &'a str) -> Self {
         Self(s.into())
+    }
+}
+
+#[cfg(feature = "client-api-proto")]
+impl From<medea_client_api_proto::MemberId> for Id {
+    fn from(id: medea_client_api_proto::MemberId) -> Self {
+        id.0.into()
     }
 }
 
@@ -213,7 +235,17 @@ pub type Sids = HashMap<Id, Sid>;
     PartialEq,
     PartialOrd,
 )]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(transparent))]
 pub struct PublicUrl(Url);
+
+impl FromStr for PublicUrl {
+    type Err = url::ParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        s.parse().map(Self)
+    }
+}
 
 /// Credentials of a [`Member`] media [`Element`] for its client side to
 /// authorize via [Client API] with.
@@ -221,6 +253,8 @@ pub struct PublicUrl(Url);
 /// [`Element`]: crate::Element
 /// [Client API]: https://tinyurl.com/266y74tf
 #[derive(Clone, Debug, Eq, From, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "lowercase"))]
 pub enum Credentials {
     /// [Argon2] hash of credentials.
     ///
@@ -267,6 +301,8 @@ impl Credentials {
     PartialEq,
     PartialOrd,
 )]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(transparent))]
 #[from(types(String))]
 #[into(owned(types(String)))]
 pub struct PlainCredentials(Box<str>); // TODO: Use `secrecy` crate.