+++ proto/control-api/src/control/room.rs
@@ -1,26 +1,49 @@
 //! [`Room`] definitions.
 
-use std::collections::HashMap;
-
 use derive_more::{AsRef, Display, From, Into};
 use ref_cast::RefCast;
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
-use super::{member, Member};
+use super::{member, Pipeline};
 
 /// Media [`Element`] representing a single space where multiple [`Member`]s can
 /// interact with each other.
 ///
 /// [`Element`]: crate::Element
 /// [`Member`]: crate::Member
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 pub struct Room {
     /// ID of this [`Room`] media [`Element`].
     ///
     /// [`Element`]: crate::Element
     pub id: Id,
 
+    /// [`Room`] spec.
+    pub spec: Spec,
+}
+
+/// [`Room`] spec.
+#[derive(Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(transparent))]
+pub struct Spec {
     /// Media pipeline representing [`Member`]s of this [`Room`].
-    pub pipeline: HashMap<member::Id, Member>,
+    ///
+    /// [`Member`]: crate::Member
+    pub spec: Pipeline<member::Id, Element>,
+}
+
+/// Possible [`Element`]s of a [`Room`]'s [`Pipeline`].
+#[derive(Clone, Debug, Eq, From, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(tag = "kind"))]
+pub enum Element {
+    /// [`Member`] media [`Element`] of the [`Room`]'s [`Pipeline`].
+    ///
+    /// [`Member`]: crate::Member
+    Member(member::Spec),
 }
 
 /// ID of a [`Room`] media [`Element`].
@@ -40,6 +63,8 @@ pub struct Room {
     PartialOrd,
     RefCast,
 )]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(transparent))]
 #[from(types(String))]
 #[into(owned(types(String)))]
 #[repr(transparent)]
@@ -49,5 +74,12 @@ pub struct Id(Box<str>);
 impl<'a> From<&'a str> for Id {
     fn from(s: &'a str) -> Self {
         Self(s.into())
+    }
+}
+
+#[cfg(feature = "client-api-proto")]
+impl From<medea_client_api_proto::RoomId> for Id {
+    fn from(id: medea_client_api_proto::RoomId) -> Self {
+        id.0.into()
     }
 }