+++ proto/control-api/CHANGELOG.md
@@ -20,11 +20,15 @@ All user visible changes to this project will be documented in this file. This p
 - channel:
     - `ControlApi` and `CallbackApi` implementations ([#65]).
 - Cargo features:
-    - `client` and `server` ([#56]).
+    - `client` and `server` ([#56]);
+    - `serde` ([#67]);
+    - `client-api-proto` ([#68]).
 
 [#55]: /../../pull/55
 [#56]: /../../pull/56
 [#65]: /../../pull/65
+[#67]: /../../pull/67
+[#68]: /../../pull/68